refactor(server): extract shutdown/restart into server_shutdown.go
Move getDoneChan, Close, RegisterOnShutdown, RegisterOnRestart,
Shutdown, Restart, startProcess, checkProcessMsg, closeDoneChanLocked
(and shutdownPollInterval) out of server.go into server_shutdown.go.
Same package, mechanical move.

server.go is now 299 lines (down from 1098); the God File (Critical #2)
is decomposed into 5 cohesive files: server.go (home/types/ctors/Serve),
server_conn.go, server_dispatch.go, server_response.go, server_shutdown.go.
