+++ server/server.go
@@ -7,8 +7,6 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"os"
-	"os/exec"
 	"regexp"
 	"strings"
 	"sync"
@@ -173,12 +171,6 @@ func (s *Server) ActiveClientConn() []net.Conn {
 	return result
 }
 
-func (s *Server) getDoneChan() <-chan struct{} {
-	return s.doneChan
-}
-
-// Serve starts and listens RPC requests.
-// It is blocked until receiving connections from clients.
 func (s *Server) Serve(network, address string) (err error) {
 	var ln net.Listener
 	ln, err = s.makeListener(network, address)
@@ -280,178 +272,6 @@ func (s *Server) ServeWS(conn *websocket.Conn) {
 
 	conn.PayloadType = websocket.BinaryFrame
 	s.serveConn(conn)
-}
-
-// Close immediately closes all active net.Listeners.
-func (s *Server) Close() error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	var err error
-	if s.ln != nil {
-		err = s.ln.Close()
-	}
-	for c := range s.activeConn {
-		c.Close()
-		delete(s.activeConn, c)
-		s.Plugins.DoPostConnClose(c)
-	}
-	s.closeDoneChanLocked()
-
-	if s.pool != nil {
-		s.pool.StopAndWaitFor(10 * time.Second)
-	}
-
-	return err
-}
-
-// RegisterOnShutdown registers a function to call on Shutdown.
-// This can be used to gracefully shutdown connections.
-func (s *Server) RegisterOnShutdown(f func(s *Server)) {
-	s.mu.Lock()
-	s.onShutdown = append(s.onShutdown, f)
-	s.mu.Unlock()
-}
-
-// RegisterOnRestart registers a function to call on Restart.
-func (s *Server) RegisterOnRestart(f func(s *Server)) {
-	s.mu.Lock()
-	s.onRestart = append(s.onRestart, f)
-	s.mu.Unlock()
-}
-
-var shutdownPollInterval = 1000 * time.Millisecond
-
-// Shutdown gracefully shuts down the server without interrupting any
-// active connections. Shutdown works by first closing the
-// listener, then closing all idle connections, and then waiting
-// indefinitely for connections to return to idle and then shut down.
-// If the provided context expires before the shutdown is complete,
-// Shutdown returns the context's error, otherwise it returns any
-// error returned from closing the Server's underlying Listener.
-func (s *Server) Shutdown(ctx context.Context) error {
-	var err error
-	if atomic.CompareAndSwapInt32(&s.inShutdown, 0, 1) {
-		log.Info("shutdown begin")
-
-		s.mu.Lock()
-
-		// 主动注销注册的服务
-		s.UnregisterAll()
-
-		if s.ln != nil {
-			s.ln.Close()
-		}
-		for conn := range s.activeConn {
-			if tcpConn, ok := conn.(*net.TCPConn); ok {
-				tcpConn.CloseRead()
-			}
-		}
-		s.mu.Unlock()
-
-		// wait all in-processing requests finish.
-		ticker := time.NewTicker(shutdownPollInterval)
-		defer ticker.Stop()
-	outer:
-		for {
-			if s.checkProcessMsg() {
-				break
-			}
-			select {
-			case <-ctx.Done():
-				err = ctx.Err()
-				break outer
-			case <-ticker.C:
-			}
-		}
-
-		s.jsonrpcHTTPServerLock.Lock()
-		if s.gatewayHTTPServer != nil {
-			if err := s.closeHTTP1APIGateway(ctx); err != nil {
-				log.Warnf("failed to close gateway: %v", err)
-			} else {
-				log.Info("closed gateway")
-			}
-		}
-		s.jsonrpcHTTPServerLock.Unlock()
-
-		if s.jsonrpcHTTPServer != nil {
-			if err := s.closeJSONRPC2(ctx); err != nil {
-				log.Warnf("failed to close JSONRPC: %v", err)
-			} else {
-				log.Info("closed JSONRPC")
-			}
-		}
-
-		s.mu.Lock()
-		for conn := range s.activeConn {
-			conn.Close()
-			delete(s.activeConn, conn)
-			s.Plugins.DoPostConnClose(conn)
-		}
-		s.closeDoneChanLocked()
-
-		s.mu.Unlock()
-
-		log.Info("shutdown end")
-
-	}
-	return err
-}
-
-// Restart restarts this server gracefully.
-// It starts a new rpcx server with the same port with SO_REUSEPORT socket option,
-// and shutdown this rpcx server gracefully.
-func (s *Server) Restart(ctx context.Context) error {
-	pid, err := s.startProcess()
-	if err != nil {
-		return err
-	}
-	log.Infof("restart a new rpcx server: %d", pid)
-
-	// TODO: is it necessary?
-	time.Sleep(3 * time.Second)
-	return s.Shutdown(ctx)
-}
-
-func (s *Server) startProcess() (int, error) {
-	argv0, err := exec.LookPath(os.Args[0])
-	if err != nil {
-		return 0, err
-	}
-
-	// Pass on the environment and replace the old count key with the new one.
-	var env []string
-	env = append(env, os.Environ()...)
-
-	originalWD, _ := os.Getwd()
-	allFiles := []*os.File{os.Stdin, os.Stdout, os.Stderr}
-	process, err := os.StartProcess(argv0, os.Args, &os.ProcAttr{
-		Dir:   originalWD,
-		Env:   env,
-		Files: allFiles,
-	})
-	if err != nil {
-		return 0, err
-	}
-	return process.Pid, nil
-}
-
-func (s *Server) checkProcessMsg() bool {
-	size := atomic.LoadInt32(&s.handlerMsgNum)
-	log.Info("need handle in-processing msg size:", size)
-	return size == 0
-}
-
-func (s *Server) closeDoneChanLocked() {
-	select {
-	case <-s.doneChan:
-		// Already closed. Don't close again.
-	default:
-		// Safe to close here. We're the only closer, guarded
-		// by s.mu.RegisterName
-		close(s.doneChan)
-	}
 }
 
 var ip4Reg = regexp.MustCompile(`^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$`)
